+++ omega_miya/onebot_api/__init__.py
@@ -1,9 +1,16 @@
 """
 @Author         : Ailitonia
 @Date           : 2022/03/30 20:27
-@FileName       : onebot_api.py
+@FileName       : __init__.py.py
 @Project        : nonebot2_miya 
-@Description    : [Deactivated]OneBot Api
+@Description    : Onebot Api
 @GitHub         : https://github.com/Ailitonia
 @Software       : PyCharm 
 """
+
+from .gocq import GoCqhttpBot
+
+
+__all__ = [
+    'GoCqhttpBot'
+]
