+++ omega_miya/onebot_api/gocq/model/__init__.py
@@ -1,7 +1,7 @@
 """
 @Author         : Ailitonia
 @Date           : 2022/04/14 21:40
-@FileName       : __init__.py
+@FileName       : __init__.py.py
 @Project        : nonebot2_miya 
 @Description    : go-cqhttp model
 @GitHub         : https://github.com/Ailitonia
@@ -11,10 +11,10 @@
 from .file import (Cookies, CSRF, Credentials, ImageFile, OcrImageResult, CanSendImage,
                    RecordFile, CanSendRecord, DownloadedFile, VersionInfo, Status, OnlineClients, UrlSafely)
 from .group import (GroupInfo, GroupHonor, GroupSystemMessage, GroupFileSystemInfo, GroupRootFiles, GroupFolderFiles,
-                    GroupFileResource, GroupAtAllRemain, GroupEssenceMessage, GroupNotice)
+                    GroupFileResource, GroupAtAllRemain, GroupEssenceMessage)
 from .guild import (GuildServiceProfile, GuildInfo, GuildMeta, ChannelInfo, GuildMemberList, GuildMemberProfile,
                     GuildRoles, CreatedGuildRoles, SentGuildMessage, ReceiveGuildMessage, TopicChannelFeedInfo)
-from .message import SentMessage, SentForwardMessage, ReceiveMessage, GroupMessageHistory, ReceiveForwardMessage
+from .message import SentMessage, ReceiveMessage, GroupMessageHistory, ReceiveForwardMessage
 from .user import StrangerInfo, LoginInfo, FriendInfo, GroupUser, Anonymous, QidianAccountUser
 
 
@@ -41,9 +41,7 @@ __all__ = [
     'GroupFileResource',
     'GroupAtAllRemain',
     'GroupEssenceMessage',
-    'GroupNotice',
     'SentMessage',
-    'SentForwardMessage',
     'ReceiveMessage',
     'GroupMessageHistory',
     'ReceiveForwardMessage',
