+++ omega_miya/onebot_api/model/__init__.py
@@ -3,12 +3,12 @@
 @Date           : 2022/04/13 21:55
 @FileName       : model.py
 @Project        : nonebot2_miya 
-@Description    : OneBot Data Model
+@Description    : Onebot Data Model
 @GitHub         : https://github.com/Ailitonia
 @Software       : PyCharm 
 """
 
-from .base_model import BaseOneBotModel
+from .base_model import BaseOnebotModel
 from .group import GroupInfo, GroupHonor
 from .message import SentMessage, ReceiveMessage, CustomNodeMessage
 from .user import LoginInfo, FriendInfo, StrangerInfo, GroupUser
@@ -16,7 +16,7 @@ from .file import ImageFile, RecordFile, CanSendImage, CanSendRecord, Cookies, C
 
 
 __all__ = [
-    'BaseOneBotModel',
+    'BaseOnebotModel',
     'GroupInfo',
     'GroupHonor',
     'SentMessage',
