+++ omega_miya/onebot_api/model/base_model.py
@@ -3,7 +3,7 @@
 @Date           : 2022/04/13 22:08
 @FileName       : base_model.py
 @Project        : nonebot2_miya 
-@Description    : OneBot Base Model
+@Description    : Onebot Base Model
 @GitHub         : https://github.com/Ailitonia
 @Software       : PyCharm 
 """
@@ -11,7 +11,7 @@
 from pydantic import BaseModel
 
 
-class BaseOneBotModel(BaseModel):
+class BaseOnebotModel(BaseModel):
     class Config:
         extra = 'ignore'
         allow_mutation = False
