--- src/service/_onebot_api/v11/__init__.py
@@ -1,16 +1,0 @@
-"""
-@Author         : Ailitonia
-@Date           : 2023/3/18 1:53
-@FileName       : v11
-@Project        : nonebot2_miya
-@Description    : onebot v11
-@GitHub         : https://github.com/Ailitonia
-@Software       : PyCharm 
-"""
-
-from .gocq import Gocqhttp
-
-
-__all__ = [
-    'Gocqhttp'
-]
